Edits for readme.md
In "Getting Started" section, changed "clone" to "cloned" to be more grammatically-correct.

In "Builds" section, rewrote sentence "Haunted comes in a few builds, pick the one depending on which environment you are using:" to "Haunted comes in a few builds. Pick one based on your chosen environment:". The information is more clearly presented with a period instead of a comma, and the second sentence is more concise.

In "Cleaning up side-effects" section, changed "if" to "of" in "An example if when you..." It's more understandable now.