Fix useEffect typing
When using `haunted` in a TS project I noticed that the `.d.ts` does not seem to have a typing for the dependencies array. At first I assumed `haunted` just didn't support one, but it seems like form looking at other issues / the code that it does? It definitely works without it so I have set it as optional in the types.