Feat/lit2 (#369)
* wip: upgrade to lit 2

* Replacing references of NodePart to ChildPart per Lit 2.0 upgrade guide.
Upgrading testing dependency to use Lit 2.0.
Five failing tests.

* WIP: Virtual components w/ Lit 2.0. Three failing tests.

* Some tweaks, but still have failing tests.

* Switching to AsyncDirective per the lit 2.0 upgrade guide.
One failing test.

* The lit 2 upgrade guide had indicated that the weakmaps should be
removed in favor of state within the directive class. In our case
it turns out we needed those weakmaps because that is how we keep
track of schedulers for the virtual components.

* Upgrading some dev dependencies in an effort to purge older versions
of lit from the lock file. Only partially successful.

* Updating check in virtual component update() method to address issue #337.

* Reverting rocket dev dependencies.

Co-authored-by: Benny Powers <web@bennypowers.com>